// src/app/models/careerModel.ts
export class careerModel{
    id?: number | undefined;
    name?: string | undefined;
    boss?: string | undefined;
    area?: string | undefined;
    institute_id?: number | undefined;
    pension?: number | undefined;
    quantityCourse?: number | undefined;
    quantitySemester?: number | undefined;
    active?: boolean | undefined;
    constructor() {}
}